Restore r332345 and r332346 from head, merged to stable/11 as part of r333006, which was reverted in r334735.
r332345 fixes makefs(8) invocation after head revision r331843,
where makefs(8) was updated to be in sync with NetBSD.

r332346 fixes the $bootable variable position so the platformid
option is correctly applied.

This is a direct commit to releng/11.2, as these two revisions
were part of a total of four revisions merge to stable/11 (at
the time) in r333006.

Approved by:	re (bdrewery)
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/releng/11.2@334839 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f