top(1): permit sub-second delay times
This removes the getuid check for delay==0. It didn't prevent users from
writing similar programs in the general case. In theory, if top(1) is
among one of the few restricted programs you're allowed to run, it may
have helped a little, but there are better ways of handling that case.


git-svn-id: https://svn.freebsd.org/base/head@334906 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f