User service foo rather than /etc/rc.d/foo.
devd predates service in the system. Modernize usage to use service to
start/stop things in reaction to events rather than calling the rc
file directly.

This was pointed out in my talk at BSDcan as well as indirectly
referrred to as a barrier to entry for OpenRC in that working group.


git-svn-id: https://svn.freebsd.org/base/head@334982 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f