MFC r309925, r309931, r309933, r310035, r310278, r310310, r310311,     r310323, r310349, r310350, r310351, r310352, r310383, r310384,     r310385, r310386, r310393, r310453, r310456, r310494, r310504,     r310528, r310890, r310893, r310974, r311918, r312921, r313357,     r314563, r314585, r314642, r315322, r315618, r315620, r315622,     r315643, r316951, r316973, r326338, r326339, r326573, r331270,     r332099, r332110, r332111, r332118, r332165, r332510 and r332511.
This commit brings syslogd(8) in sync with the copy in HEAD. The key
improvement of this change is that it adds support for RFC 5424 log
ingestion and exposition (enabled by passing in -O rfc5424). This allows
for saner logging in environments with multiple time zones.

The list of changes to merge back were obtained by running:

    svn mergeinfo --show-revs eligible \
        ^/head/usr.sbin/syslogd ^/stable/11/usr.sbin/syslogd

Of the commits listed, r314436, r325188 and r326025 were excluded, as
they affect a significant number of unrelated files (SPDX and 4-clause
license renumbering). Due to the large number of directly committed
changes on this branch, I had no choice but to perform the merge as
follows:

    svn merge --accept=theirs-full -c <list of revisions> ^/head .

This would, however, cause some unrelated changes, such as undoing the
r333356 (MFC of r332877) and still adding the SPDX tag to syslogd.c.
These have been reverted manually.

Requested by:	Dave Cottlehuber
Thanks to:	dim@ for sharing his insight on hackers@


git-svn-id: https://svn.freebsd.org/base/stable/11@335059 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f