rk_pll: Add support for mode
RockChip PLL have two modes controlled by a register, a "slow mode" (the
default one) where the frequency is derived from the 24Mhz oscillator on the
board, and a "normal" one when the pll take it's input from the real PLL output.

Default the mode to normal for all the PLLs.


git-svn-id: https://svn.freebsd.org/base/head@335108 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f