MFC r334948:
Fix build of bxe with base gcc on i386

Casting from rman_res_t to a pointer results in "cast to pointer from
integer of different size" warnings with base gcc on i386, so print
these without casting.  The kva field of struct bxe_bar is of type
vm_offset_t, which can be 32 or 64 bit, so cast it to uintmax_t before
printing.

Reviewed by:	markj
Differential Revision: https://reviews.freebsd.org/D15733


git-svn-id: https://svn.freebsd.org/base/stable/11@335289 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f