if_rk_dwc: Fix delays handling
The property are named {t,r}x_delay and not {t,r}-delay.
The upper bits of the register are a mask of which bits is allowed
to be written, set it otherwise we write nothing.
OF_getencprop returns <0 = for an error.

Pointy Hat: myself
Reported by:	jmcneill (delay and mask bits)


git-svn-id: https://svn.freebsd.org/base/head@335443 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f