MFC r335296:
Fix build of aibs with base gcc on i386

Add a few intermediate casts to intptr_t to suppress "cast to pointer
from integer of different size" warnings from gcc.  In this case, the
'arg2' part of SYSCTL_HANDLER_ARGS is used to pass in a pointer, via an
intermediate intmax_t, so no information is lost.

Reviewed by:	avg
Differential Revision: https://reviews.freebsd.org/D15725


git-svn-id: https://svn.freebsd.org/base/stable/11@335471 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f