MFC r335297:
Fix build of ndis with base gcc on i386

Casting from rman_res_t to a pointer results in "cast to pointer from
integer of different size" warnings with base gcc on i386, so use an
intermediate cast to uintptr_t to suppress it.  In this case, the I/O
port range is effectively limited to the range of 0..65535.

Reviewed by:	imp
Differential Revision: https://reviews.freebsd.org/D15746


git-svn-id: https://svn.freebsd.org/base/stable/11@335472 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f