Revert preference to be an int.
While in base we use it as a boolean (of the wrong spelling), there's
at least one out of tree user that needs it to be int since priorirty
is a small int, not a 0/1. In deference to the time it's wasted me and
my team, push this up into FreeBSD for whatever short life boot1 may
have in the tree.


git-svn-id: https://svn.freebsd.org/base/head@335773 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f