Document the "#mds_path" suffix for the "-p" command line option.
r335871 added support for an optional suffix of "#mds_path" that can be
applied to each entry in the "-p" option argument. This specifies that
the DS should be used to store files for the file system on the MDS
at "mds_path".
This patch documents this optional suffix.

This is a content change.


git-svn-id: https://svn.freebsd.org/base/head@335872 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f