mxge: fix panic at module unload
r333175 (multicast changes) exposed a bug where
mxge was not checking to see if the driver was being
unloaded while handing ioctls that touch hardware.
As a result, now that in6m_disconnect() is run from
an async gtaskq, it was busy-waiting in mxge_send_cmd()
while the mcast list was destroyed.


git-svn-id: https://svn.freebsd.org/base/head@335957 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f