Remove gateworks nanobsd support.
A quick test of this shows multiple problems. Rather than fix the
problems, just retire this board's support. It's for a 12 year old
board that's been out of production for at least 7 years and generally
lacks the memory to run even a stripped down NanoBSD image well. It's
not really relevant anymore.


git-svn-id: https://svn.freebsd.org/base/head@336136 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f