MFC r335886:
pfctl: Don't retrieve interface list if '-n' is set

If '-n' is set we don't use the list of skip interfaces, so don't retrieve it.
This fixes issues if 'pfctl -n' is used before the pf module is loaded. This
was broken by r333181.

Reported by:	Jakub Chromy <hicks AT cgi.cz>


git-svn-id: https://svn.freebsd.org/base/stable/11@336164 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f